+++ src/index.ts
@@ -1,2 +1,12 @@
 // reexport as named-object for Node
 export * as DsUtil from "./dist";
+// reexport all types and interfaces that could be used by the user
+export * from "./base/types/DsGrammarGeneric.type";
+export * from "./base/types/LanguageTaggedString.type";
+export * from "./base/types/NodeSchema.type";
+export * from "./v5/types/DataTypesV5.type";
+export * from "./v5/types/DsGrammarV5.type";
+export * from "./v7/types/DataTypesV7.type";
+export * from "./v7/types/DsGrammarV7.type";
+export * from "./v7/types/PathGrammarV7.type";
+export * from "./v7/types/VerificationV7.type";